fix: skip manifest validation in pre-commit if a git rebase is in progress
Currently, manifest validation is initiated in the post-commit git-hook
using the pre-commit tool. Unfortunately, the post-commit git-hook
cannot be easily bypassed with the --no-verify option and runs during
each commit in a rebase, which can be annoying and time-consuming for
larger rebases. The process can be sped up by skipping the check if a
rebase is in progress. This introduces a new option --skip-on-rebase,
which is used only when manifest validation is started through
pre-commit and skips the validation. This makes the process about five
times faster, though still slower than when the hook is completely
disabled due to the time required to spawn pre-commit and esp-idf-sbom.

If needed, the post-commit hook can be temporarily disabled with:

pre-commit uninstall -t pre-commit or chmod -x .git/hooks/post-commit

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>